requirements.txt
pydantic
python-dotenv
ruff
openai==2.6.0
pillow